+++ packages/e2e/react-router/v7/app/routes/repro-702.tsx
@@ -1,0 +1,3 @@
+import { Repro702 } from 'e2e-shared/specs/repro-702'
+
+export default Repro702
