+++ packages/e2e/react-router/v7/cypress/e2e/shared/repro-702.cy.ts
@@ -1,0 +1,5 @@
+import { testRepro702 } from 'e2e-shared/specs/repro-702.cy'
+
+testRepro702({
+  path: '/repro-702'
+})
