+++ packages/e2e/shared/specs/repro-702.cy.ts
@@ -1,8 +1,9 @@
-/// <reference types="cypress" />
+import { createTest } from '../create-test'
 
-describe('repro-702', () => {
+export const testRepro702 = createTest('repro-702', ({ path }) => {
   it('mounts components with the correct state on load', () => {
-    cy.visit('/app/repro-702?a=test&b=test')
+    cy.visit(`${path}?a=test&b=test`)
+    cy.contains('#hydration-marker', 'hydrated').should('be.hidden')
     cy.get('#conditional-a-useQueryState').should('have.text', 'test pass')
     cy.get('#conditional-a-useQueryStates').should('have.text', 'test pass')
     cy.get('#conditional-b-useQueryState').should('have.text', 'test pass')
@@ -10,13 +11,11 @@ describe('repro-702', () => {
   })
 
   it('mounts components with the correct state after an update', () => {
-    cy.visit('/app/repro-702')
+    cy.visit(path)
     cy.contains('#hydration-marker', 'hydrated').should('be.hidden')
-
     cy.get('#trigger-a').click()
     cy.get('#conditional-a-useQueryState').should('have.text', 'test pass')
     cy.get('#conditional-a-useQueryStates').should('have.text', 'test pass')
-
     cy.get('#trigger-b').click()
     cy.get('#conditional-b-useQueryState').should('have.text', 'test pass')
     cy.get('#conditional-b-useQueryStates').should('have.text', 'test pass')
