+++ packages/nuqs/src/useQueryStates.test.ts
@@ -607,13 +607,12 @@ describe('useQueryStates: update sequencing', () => {
     expect(p1).toBeInstanceOf(Promise)
     expect(p2).toBeInstanceOf(Promise)
     expect(p1).not.toBe(p2)
-    // Note: our mock adapter does not save search params, so there is no merge
-    await expect(p1).resolves.toEqual(new URLSearchParams('?a=a'))
     await expect(p2).resolves.toEqual(new URLSearchParams('?b=b'))
+    await expect(p1).resolves.toEqual(new URLSearchParams('?b=b&a=a'))
     expect(onUrlUpdate).toHaveBeenCalledTimes(2)
     // b updates first, then a
     expect(onUrlUpdate.mock.calls[0]![0].queryString).toEqual('?b=b')
-    expect(onUrlUpdate.mock.calls[1]![0].queryString).toEqual('?a=a')
+    expect(onUrlUpdate.mock.calls[1]![0].queryString).toEqual('?b=b&a=a')
   })
 
   it('aborts a debounced update when pushing a throttled one', async () => {
@@ -643,7 +642,6 @@ describe('useQueryStates: update sequencing', () => {
     expect(p1).toBeInstanceOf(Promise)
     expect(p2).toBeInstanceOf(Promise)
     expect(p1).not.toBe(p2)
-    // Note: our mock adapter does not save search params, so there is no merge
     await expect(p1).resolves.toEqual(new URLSearchParams('?test=pass'))
     await expect(p2).resolves.toEqual(new URLSearchParams('?test=pass'))
     expect(onUrlUpdate).toHaveBeenCalledTimes(1)
@@ -675,11 +673,16 @@ describe('useQueryStates: update sequencing', () => {
     expect(p1).toBeInstanceOf(Promise)
     expect(p2).toBeInstanceOf(Promise)
     expect(p1).not.toBe(p2)
-    // Note: our mock adapter does not save search params, so there is no merge
-    await expect(p1).resolves.toEqual(new URLSearchParams('?a=debounced'))
+    console.log((await p1!).toString())
+    console.log((await p2!).toString())
     await expect(p2).resolves.toEqual(new URLSearchParams('?b=pass'))
+    await expect(p1).resolves.toEqual(
+      new URLSearchParams('?b=pass&a=debounced')
+    )
     expect(onUrlUpdate).toHaveBeenCalledTimes(2)
     expect(onUrlUpdate.mock.calls[0]![0].queryString).toEqual('?b=pass')
-    expect(onUrlUpdate.mock.calls[1]![0].queryString).toEqual('?a=debounced')
+    expect(onUrlUpdate.mock.calls[1]![0].queryString).toEqual(
+      '?b=pass&a=debounced'
+    )
   })
 })
